feat: versao aula 11/6
